Moving components to components section
