+++ pages/about.js
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import styles from '../styles/Home.module.css'
-import Header from './Header'
-import Footer from './Footer'
+import Header from '../components/Header'
+import Footer from '../components/Footer'
 import { MdCode,MdDownload,MdPerson } from "react-icons/md";
 
 export default function About() {
