+++ pages/install.js
@@ -3,8 +3,8 @@ import styles from "../styles/Home.module.css";
 import { WindowMockup } from "react-mockup";
 import "react-mockup/dist/index.css";
 import { MdCode, MdDownload, MdPerson } from "react-icons/md";
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 
 export default function Guide() {
   return (
