+++ pages/opensource.js
@@ -4,8 +4,8 @@ import Head from "next/head";
 import styles from "../styles/Home.module.css";
 import "react-mockup/dist/index.css";
 import { MdCode, MdOutlineDownload, MdLogin } from "react-icons/md";
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 import { Widget, PopupButton } from "@typeform/embed-react";
 
 export default function Guide() {
