+++ pages/x86_64.js
@@ -1,6 +1,6 @@
 import Head from 'next/head'
 import styles from '../styles/Home.module.css'
-import Header from './Header'
+import Header from '../components/Header'
 
 export default function About() {
     return (
